// src/admin-subscriptions/subscriptions-designe-settings/dto/subscriptions-designe-settings-response.dto.ts
export class SubscriptionsDesigneSettingsResponseDto {
  subscriptionsDesigneSettingId: string | undefined;
  url: string | undefined;
  brandOne: string | undefined;
  brandTwo: string | undefined;
  brandThree: string | undefined;
  brandFour: string | undefined;
  primaryColor: string | undefined;
  secondaryColor: string | undefined;
  baseColor: string | undefined;
  infoColor: string | undefined;
  warningColor: string | undefined;
  successColor: string | undefined;
  errorColor: string | undefined;
  lightColor: string | undefined;
  darkColor: string | undefined;
  letterFont: string | undefined;
}